api/requirements.txt
fastapi==0.63.0
kafka-python==2.0.2
pymongo[srv]==3.11.2
sentry-sdk==1.0.0
requests
wait-for-it==2.2.0
pyjwt==2.1.0
pycryptodomex==3.10.1
